+++ src/main/kotlin/com/travel/travelapi/controllers/SourcePlaceController.kt
@@ -1,12 +1,13 @@
 package com.travel.travelapi.controllers
 
-import com.travel.travelapi.models.Category
-import com.travel.travelapi.models.CategoryPlace
+import com.travel.travelapi.auth.TravelUserDetails
+import com.travel.travelapi.exceptions.UnauthorizedException
 import com.travel.travelapi.models.Source
 import com.travel.travelapi.models.SourcePlace
+import com.travel.travelapi.services.PlaceService
 import com.travel.travelapi.services.SourcePlaceService
-import com.travel.travelapi.services.SourceService
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RequestParam
@@ -15,7 +16,8 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/sourceplace")
-class SourcePlaceController(@Autowired private val sourcePlaceService: SourcePlaceService) {
+class SourcePlaceController(@Autowired private val sourcePlaceService: SourcePlaceService,
+                            @Autowired private val placeService: PlaceService) {
 
 
     /**
@@ -31,7 +33,16 @@ class SourcePlaceController(@Autowired private val sourcePlaceService: SourcePla
      */
     @RequestMapping("/insert")
     fun insertSource(@RequestBody categoryPlace: List<SourcePlace>){
+
+        //Getting the authenticated user
+        val principal = SecurityContextHolder.getContext().authentication.principal as TravelUserDetails
+
         for(c: SourcePlace in categoryPlace){
+
+            //Checking if user has access to modify this place's sources. If not, this line will throw an exception
+            checkModifyAccess(principal, c.fk_placeId!!)
+
+            //Inserting source for this place
             sourcePlaceService.insertSource(c)
         }
     }
@@ -40,21 +51,59 @@ class SourcePlaceController(@Autowired private val sourcePlaceService: SourcePla
      * Delete source from a place
      */
     @RequestMapping("/delete")
-    fun deleteSource(@RequestBody categoryPlace: List<SourcePlace>){
-        for(c: SourcePlace in categoryPlace){
+    fun deleteSource(@RequestBody sourcePlaces: List<SourcePlace>){
+
+        //Getting the authenticated user
+        val principal = SecurityContextHolder.getContext().authentication.principal as TravelUserDetails
+
+        for(c: SourcePlace in sourcePlaces){
+
+            //Checking if user has access to modify this place's sources. If not, this line will throw an exception
+            checkModifyAccess(principal, c.fk_placeId!!)
+
             sourcePlaceService.deleteSource(c)
         }
+    }
+
+    /**
+     * Checks if user has permission to modify sources of a place with the given ID.
+     * Throws an exception if user is not permitted
+     */
+    @Throws(UnauthorizedException::class)
+    fun checkModifyAccess(user: TravelUserDetails, placeId: Int){
+
+        //Getting the relevant parking from database
+        val place = placeService.selectById(placeId)
+
+        //Checking if user has permission to modify all parkings unconditionally
+        if(user.hasAuthority("source:modify_unrestricted")){
+            return
+            //Else we check if user has authority to modify their parking(s) and the given parking is created by them
+        }else if(place.userId != null
+                && place.userId!!.toLong() == user.id
+                && !place.isVerified!!
+                && !place.isPublic!!
+                && user.hasAuthority("source:modify"))   {
+            return
+        }else throw UnauthorizedException("Insufficient authority") //If all else fails, we throw an exception
     }
 
     /**
      * Update place sources
      */
     @RequestMapping("/update")
-    fun updateTagForPlace(@RequestBody catPlaces: List<Source>, @RequestParam(name="p") id: Int){
+    fun updateSourceForPlace(@RequestBody sources: List<Source>, @RequestParam(name="p") id: Int){
+
+        //Getting the authenticated user
+        val principal = SecurityContextHolder.getContext().authentication.principal as TravelUserDetails
+
+        //Checking if user has access to modify this place's sources. If not, this line will throw an exception
+        checkModifyAccess(principal, id)
+
         sourcePlaceService.deleteSourcesById(id)
 
         val cats = ArrayList<SourcePlace>()
-        for(t: Source in catPlaces)
+        for(t: Source in sources)
             cats.add(SourcePlace(t.sourceId!!, id))
 
         insertSource(cats)
